+++ src/UnitySampleProject/Assets/Samples/web3.unity SDK/2.5.0/Web3.Unity Samples/Scripts/Scenes/SampleMain/Marketplace/MarketplaceCalls.cs
@@ -264,7 +264,7 @@ public class MarketplaceCalls : MonoBehaviour
     /// </summary>
     public async void PurchaseNftFromMarketplace()
     {
-        var data = await Marketplace.PurchaseNft(marketplaceContractToBuyFrom,tokenIdToBuy, weiPriceToBuy);
+        var data = await Marketplace.PurchaseNft(marketplaceContractToBuyFrom, tokenIdToBuy, weiPriceToBuy);
         var response = SampleOutputUtil.BuildOutputValue(data);
         Debug.Log($"TX: {response}");
     }
